+++ packages/core-js/modules/es.aggregate-error.js
@@ -2,6 +2,7 @@
 var $ = require('../internals/export');
 var getPrototypeOf = require('../internals/object-get-prototype-of');
 var setPrototypeOf = require('../internals/object-set-prototype-of');
+var copyConstructorProperties = require('../internals/copy-constructor-properties');
 var create = require('../internals/object-create');
 var createNonEnumerableProperty = require('../internals/create-non-enumerable-property');
 var createPropertyDescriptor = require('../internals/create-property-descriptor');
@@ -24,6 +25,9 @@ var $AggregateError = function AggregateError(errors, message /* , options */) {
   createNonEnumerableProperty(that, 'errors', errorsArray);
   return that;
 };
+
+if (setPrototypeOf) setPrototypeOf($AggregateError, Error);
+else copyConstructorProperties($AggregateError, Error);
 
 $AggregateError.prototype = create(Error.prototype, {
   constructor: createPropertyDescriptor(1, $AggregateError),